+++ libs/common/src/main/java20/org/opensearch/common/round/RoundableFactory.java
@@ -10,6 +10,9 @@ package org.opensearch.common.round;
 
 import org.opensearch.common.annotation.InternalApi;
 
+import jdk.incubator.vector.LongVector;
+import jdk.incubator.vector.VectorSpecies;
+
 /**
  * Factory class to create and return the fastest implementation of {@link Roundable}.
  *
@@ -24,6 +27,17 @@ public final class RoundableFactory {
      */
     private static final int LINEAR_SEARCH_MAX_SIZE = 64;
 
+    /**
+     * The preferred LongVector species with the maximal bit-size supported on this platform.
+     */
+    private static final VectorSpecies<Long> LONG_VECTOR_SPECIES = LongVector.SPECIES_PREFERRED;
+
+    /**
+     * Indicates whether the vectorized (SIMD) B-tree search implementation is supported.
+     * This is true when the platform has a minimum of 4 long vector lanes.
+     */
+    private static final boolean IS_BTREE_SEARCH_SUPPORTED = LONG_VECTOR_SPECIES.length() >= 4;
+
     private RoundableFactory() {}
 
     /**
@@ -32,6 +46,8 @@ public final class RoundableFactory {
     public static Roundable create(long[] values, int size, boolean useSimdIfAvailable) {
         if (size <= LINEAR_SEARCH_MAX_SIZE) {
             return new BidirectionalLinearSearcher(values, size);
+        } else if (IS_BTREE_SEARCH_SUPPORTED && useSimdIfAvailable) {
+            return new BtreeSearcher(values, size, LONG_VECTOR_SPECIES);
         } else {
             return new BinarySearcher(values, size);
         }